corrige geracao de formulas matematicas e adiciona novas opcoes de geracao

+++ src/geg_toanki/cli.py
@@ -15,32 +15,50 @@ app = typer.Typer()
 @app.command()
 def create(
     input_path_str: str,
-    deck: str = typer.Option(..., "--deck=", "-d", help="Nome do deck no Anki"),
+    deck: str = typer.Option(..., "--deck", "-d", help="Nome do deck no Anki"),
     subdecks: bool = typer.Option(
         None,
         "--subdecks",
         "-s",
-        help="Subdecks, filhos do deck principal, gerados de acordo com a estrutura de diretórios",
+        help="Gera subdecks, filhos do deck principal, gerados de acordo com a estrutura de diretórios",
     ),
-    generate_md: bool = typer.Option(
-        False, "--generate-md", "-m", help="Gera arquivos intermediários em Markdown"
+    generate_md: str = typer.Option(
+        None,
+        "--generate-md",
+        "-m",
+        help="Gera arquivos intermediários em Markdown no diretório indicado",
+    ),
+    raw: bool = typer.Option(
+        None, "--raw", "-r", help="Processa o material diretamente sem gerar resumos"
+    ),
+    ai_summarize: bool = typer.Option(
+        None, "--ai-summary", "-a", help="Gerar resumo do arquivo com I.A."
     ),
 ):
-    # file_type = detect_file_type(file_path)
+    if not raw:
+        if generate_md:
+            # cria diretório de saída se não existir
+            output_path = Path(generate_md)
+            output_path.mkdir(parents=True, exist_ok=True)
+        else:
+            # cria diretório temporário para arquivos intermediários
+            output_path = Path("./.geg/toanki/temp")
+            output_path.mkdir(parents=True, exist_ok=True)
 
-    output_path = Path("./.geg/toanki/temp")
-    output_path.mkdir(parents=True, exist_ok=True)
-
-    input_path = Path(input_path_str)
-
-    # converte as pastas em locais temporarios
-    convert_to_md(
-        file_path_str=input_path_str,
-        output_path_str="./.geg/toanki/temp",
-        summarize=False,
-        complete=False,
-        use_ai=False,
-    )
+        # converte arquivos para Markdown
+        convert_to_md(
+            file_path_str=input_path_str,
+            output_path_str=str(output_path),
+            summarize=True,
+            complete=True,
+            use_ai=ai_summarize,
+        )
+    else:
+        output_path = Path(input_path_str)
+        print(
+            "Processando arquivos diretamente de:",
+            [f for f in output_path.rglob("*.md")],
+        )
 
     print("Arquivos convertidos para Markdown...")
 
@@ -73,7 +91,8 @@ def create(
 
     # move temporarios para /processed
     # mudar apenas para output_path para depois poder guardar histórico em .geg/anki
-    shutil.rmtree(".geg/toanki")
+    if not raw and not generate_md:
+        shutil.rmtree(".geg/toanki")
 
 
 app()
